KNXex.GroupAddress (knxex v0.1.0)

KNX group address.

Types

t()

@type t() :: %KNXex.GroupAddress{
  main: non_neg_integer(),
  middle: non_neg_integer(),
  sub: non_neg_integer()
}

Represents a KNX group address.

Functions

from_raw_address(address)

@spec from_raw_address(non_neg_integer()) :: t()

Parses the raw group address (16bit integer) into a GroupAddress struct.

from_string(group)

@spec from_string(String.t()) :: {:ok, t()} | {:error, :invalid_binary}

Parses a group address from a string into a GroupAddress struct.

make(main, middle, sub)

@spec make(non_neg_integer(), non_neg_integer(), non_neg_integer()) :: t()

Creates a GroupAddress struct from the given main, middle and sub group addresses parts.

to_raw_address(group)

@spec to_raw_address(t()) :: non_neg_integer()

Returns the raw group address (16bit integer).

to_string(group)

@spec to_string(t()) :: String.t()

Returns the group address as a string.