KNXex.IndividualAddress (knxex v0.1.0)

KNX individual address (also known as physical address).

Types

t()

@type t() :: %KNXex.IndividualAddress{
  area: non_neg_integer(),
  device: non_neg_integer(),
  line: non_neg_integer()
}

Represents a KNX individual address.

Functions

from_raw_address(address)

@spec from_raw_address(non_neg_integer()) :: t()

Parses the raw individual address (16bit integer) into a IndividualAddress struct.

from_string(individual)

@spec from_string(String.t()) :: {:ok, t()} | {:error, :invalid_binary}

Parses a individual address from a string into a IndividualAddress struct.

make(area, line, device)

@spec make(non_neg_integer(), non_neg_integer(), non_neg_integer()) :: t()

Creates a IndividualAddress struct from the given area, line and device individual addresses parts.

to_raw_address(individual)

@spec to_raw_address(t()) :: non_neg_integer()

Returns the raw individual address (16bit integer).

to_string(individual)

@spec to_string(t()) :: String.t()

Returns the individual address as a string.