KNXex.PID.DeviceCapabilities (knxex v0.1.0)

KNXnet/IP Parameter Object "Device Capabilities".

Functions

assert_name(atom1, atom2)

by_name(atom1, atom2)

by_value(atom, arg2)

has_device_management(value)
(macro)

Checks if the given device capabilities has the device management capability bit set.

has_object_server(value)
(macro)

Checks if the given device capabilities has the object server capability bit set.

has_remote_conf_and_diag(value)
(macro)

Checks if the given device capabilities has the remote configuration and diagnosis capability bit set.

has_remote_logging(value)
(macro)

Checks if the given device capabilities has the remote logging capability bit set.

has_routing(value)
(macro)

Checks if the given device capabilities has the routing capability bit set.

has_tunneling(value)
(macro)

Checks if the given device capabilities has the tunneling capability bit set.

macro_assert_name(atom1, atom2)
(macro)

macro_by_name(atom1, atom2)
(macro)

macro_by_value(atom, arg2)
(macro)